refactor(learn): centralize ATLAS_LEARN_* knobs behind one learn-settings resolver (#3722)
All 7 ATLAS_LEARN_* tuning knobs were read at scattered call sites in
pattern-cache.ts (workspace scope, getSettingAuto) and
promote-decay-scheduler.ts (platform scope, getSetting), with defaults and the
workspace-vs-platform scope policy duplicated across both files and a
cross-module DEFAULT_LATENCY_BUDGET_MS import just to share a constant.

Introduce lib/learn/learn-settings.ts as the single reader seam: it owns every
literal getSetting/getSettingAuto call, all defaults, and one consolidated doc
comment stating which knob is read at which scope and why — including the two
intentionally dual-scoped keys (CONFIDENCE_THRESHOLD, LATENCY_BUDGET_MS), which
are kept dual-scope by design, not collapsed.

pattern-cache.ts and promote-decay-scheduler.ts now delegate to the resolver and
re-export the prior public names, so agent.ts, layers.ts, and the existing test
mocks keep importing from the same modules unchanged — no behavior change.

Reads stay literal one-liners to satisfy the #3382 reader guard
(check-settings-readers.sh, green: 46 keys). No cross-module constant import.

Deferred from v0.0.17 "Performance-aware Atlas" (Architecture Backlog).
